fix(zig): fix non-existent reference to zig@v2
The action references `uses: tree-sitter/parser-test-action/zig@v2` but this file does not exist with the v2 tag.
This results in CI failure if you try to use the `parser-test-action`, even if you don't try to use the new zig
functionality.

Fix the reference to use `master` which is valid.
Fix README typo of `zig-version`.